Abstract the CONUS_Downloader logic into an AbstractGitLabDownloader
